requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
tenacity>=8.2.0
certifi>=2024.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
rich>=13.7.0
colorama>=0.4.6
tqdm>=4.66.0
markdown>=3.5.0
pdfkit>=1.0.0
shodan>=1.30.0
python-magic-bin>=0.4.14 ; platform_system=="Windows"
python-magic>=0.4.27 ; platform_system!="Windows"